When a editor is hid in ListBarWidgetImpl, we would like to generate a event (PlaceHidEvent). This should work like Focus/Maximized events.
We have to change ActivityManagerImpl, because when the part is associated with PathPlaceRequest, the place identifier was null (making impossible to
identify which place was hidden).

So in order to do that, when an activity is resolved by ActivityManagerImpl, I've made a lookup on resolvePlaceIdentifierForPathPlaceRequest method.

This feature request was generated by the need for hide the docks when Data Modeller was put in the background.

(cherry picked from commit fc41dbe45fb6a6b5dfaeb1a31acd16025b11900f)