Fix initialization ordering problem in GridLayerRedrawManager
The comparator was initialized after the singleton instance was created,
causing a null comparator to be passed to the constructor of the commands
TreeSet and hence natural ordering to be used instead which didn't work
as Commands are not Comparable.

Since this type is a singleton it shouldn't have any other static fields
than instance to guarantee that the instance is in a proper state when
created.

This also caused ClassCastExceptions with generateJsInteropExports=true.