[RHBPMS-4669] filter out invalid (null) login principals (#715)
In some cases (like one the described in the JIRA), the list of
principals contained a principal with 'null' name, which is
probably wrong. This caused a NPE further down as the name was used
to compute hashcode. Filtering the invalid principals seems to be
the easiest way to workaround that issue.